CHANGES: - Removed MESSAGE_BAN from lua - Exit code of -2 now means that run.cmd does not loop anymore.
BUGFIX:
- Creditsystem: Error on process achievements
- GameMaster: multiple coalition system errors (re-migration needed!)
- Music: error on skip if no song was playing

+++ plugins/music/sink/srs.py
@@ -56,7 +56,7 @@ class SRSSink(Sink):
             self.current = None
 
     async def skip(self) -> None:
-        if self.process.returncode is None:
+        if self.process is not None and self.process.returncode is None:
             self.process.kill()
             self.current = None
